1a/requirements.txt
opencv-python-headless
pytesseract
Pillow
pdf2image
ultralytics==8.0.192
numpy==1.26.4



